Move to scstyles.h all the prototypes of non-static scstyles.c functions
