+++ fontforge/Makefile.am
@@ -74,6 +74,7 @@ noinst_HEADERS = \
 	parsettfvar.h \
 	psread.h \
 	savefont.h \
+	scstyles.h \
 	fontforgeui.h sftextfieldP.h configure-fontforge.h views.h 	\
 	flaglist.h strlist.h charview_private.h cvundoes.h tables.h
