+++ fontforge/scripting.c
@@ -60,6 +60,7 @@
 #include "parsettf.h"
 #include "print.h"
 #include "savefont.h"
+#include "scstyles.h"
 #include <gfile.h>
 #include <utype.h>
 #include <ustring.h>
